Move dev test classes to /src/test/
