+++ server_extensions_extra/src/test/java/com/generalbytes/batm/server/extensions/extra/nano/dev/TestPaymentSupport.java
@@ -1,4 +1,4 @@
-package com.generalbytes.batm.server.extensions.extra.nano.test;
+package com.generalbytes.batm.server.extensions.extra.nano.dev;
 
 import com.generalbytes.batm.common.currencies.CryptoCurrency;
 import com.generalbytes.batm.server.extensions.TestExtensionContext;
@@ -11,6 +11,7 @@ import com.generalbytes.batm.server.extensions.extra.nano.rpc.NanoWsClient;
 import com.generalbytes.batm.server.extensions.extra.nano.wallet.node.NanoNodeWallet;
 import com.generalbytes.batm.server.extensions.payment.IPaymentRequestListener;
 import com.generalbytes.batm.server.extensions.payment.PaymentRequest;
+import org.junit.Ignore;
 
 import java.math.BigDecimal;
 import java.net.URI;
@@ -20,6 +21,7 @@ import java.net.URL;
  * THIS CLASS MAY BE IGNORED.
  * Its only purpose is to test and help during development.
  */
+@Ignore
 public class TestPaymentSupport {
 
     public static void main(String[] args) throws Exception {
